Changes to allow the program to compile!     visc_Eij and visc_Sij seem to have been taken out. And, the code no longer needs to resize them.     Also took out some warning messages

+++ cantera18/changes/liquidTransportDevelop/Cantera/src/transport/LiquidTransportData.cpp
@@ -295,7 +295,7 @@ namespace Cantera {
     doublereal t = m_thermo->temperature();
     if (t != m_temp) {    
       double tempN = 1.0;
-      for ( int i = 0; i < m_coeffs.size() ; i++ ) {
+      for (int i = 0; i < (int) m_coeffs.size() ; i++) {
 	m_prop += m_coeffs[i] * tempN;
 	tempN *= m_temp;
       }
